Move specific dynamical friction tests to their own file
