+++ tests/test_potential.py
@@ -5,7 +5,6 @@ import sys
 PY3= sys.version > '3'
 import pytest
 import numpy
-from galpy.util.bovy_conversion import velocity_in_kpcGyr
 try:
     import pynbody
     _PYNBODY_LOADED= True
@@ -3025,138 +3024,6 @@ def test_dissipative_noVelocityError():
         dummy= potential.evaluatezforces([lp,cdfc],R,z,phi=phi)
     with pytest.raises(potential.PotentialError) as excinfo:
         dummy= potential.evaluaterforces([lp,cdfc],R,z,phi=phi)
-    return None
-
-def test_ChandrasekharDynamicalFrictionForce_constLambda():
-    # Test that the ChandrasekharDynamicalFrictionForce with constant Lambda
-    # agrees with analytical solutions for circular orbits:
-    # assuming that a mass remains on a circular orbit in an isothermal halo 
-    # with velocity dispersion sigma and for constant Lambda:
-    # r_final^2 - r_initial^2 = -0.604 ln(Lambda) GM/sigma t 
-    # (e.g., B&T08, p. 648)
-    from galpy.util import bovy_conversion
-    from galpy.orbit import Orbit
-    ro,vo= 8.,220.
-    # Parameters
-    GMs= 10.**9./bovy_conversion.mass_in_msol(vo,ro)
-    const_lnLambda= 7.
-    r_init= 2.
-    dt= 2./bovy_conversion.time_in_Gyr(vo,ro)
-    # Compute
-    lp= potential.LogarithmicHaloPotential(normalize=1.,q=1.)
-    cdfc= potential.ChandrasekharDynamicalFrictionForce(\
-        GMs=GMs,const_lnLambda=const_lnLambda,
-        dens=lp) # don't provide sigmar, so it gets computed using galpy.df.jeans
-    o= Orbit([r_init,0.,1.,0.,0.,0.])
-    ts= numpy.linspace(0.,dt,1001)
-    o.integrate(ts,[lp,cdfc],method='odeint')
-    r_pred= numpy.sqrt(o.r()**2.-0.604*const_lnLambda*GMs*numpy.sqrt(2.)*dt)
-    assert numpy.fabs(r_pred-o.r(ts[-1])) < 0.01, 'ChandrasekharDynamicalFrictionForce with constant lnLambda for circular orbits does not agree with analytical prediction'
-    return None
-
-def test_ChandrasekharDynamicalFrictionForce_varLambda():
-    # Test that dynamical friction with variable Lambda for small r ranges 
-    # gives ~ the same result as using a constant Lambda that is the mean of
-    # the variable lambda
-    # Also tests that giving an axisymmetric list of potentials for the 
-    # density works
-    from galpy.util import bovy_conversion
-    from galpy.orbit import Orbit
-    ro,vo= 8.,220.
-    # Parameters
-    GMs= 10.**9./bovy_conversion.mass_in_msol(vo,ro)
-    r_init= 3.
-    dt= 2./bovy_conversion.time_in_Gyr(vo,ro)
-    # Compute evolution with variable ln Lambda
-    cdf= potential.ChandrasekharDynamicalFrictionForce(\
-        GMs=GMs,rhm=0.125,
-        dens=potential.MWPotential2014,sigmar=lambda r: 1./numpy.sqrt(2.))
-    o= Orbit([r_init,0.,1.,0.,0.,0.])
-    ts= numpy.linspace(0.,dt,1001)
-    o.integrate(ts,[potential.MWPotential2014,cdf],method='odeint')
-    lnLs= numpy.array([cdf.lnLambda(r,v) for (r,v) in zip(o.r(ts),numpy.sqrt(o.vx(ts)**2.+o.vy(ts)**2.+o.vz(ts)**2.))])
-    cdfc= potential.ChandrasekharDynamicalFrictionForce(\
-        GMs=GMs,rhm=0.125,const_lnLambda=numpy.mean(lnLs),
-        dens=potential.MWPotential2014,sigmar=lambda r: 1./numpy.sqrt(2.))
-    oc= o()
-    oc.integrate(ts,[potential.MWPotential2014,cdfc],method='odeint')
-    assert numpy.fabs(oc.r(ts[-1])-o.r(ts[-1])) < 0.05, 'ChandrasekharDynamicalFrictionForce with variable lnLambda for a short radial range is not close to the calculation using a constant lnLambda'
-    return None
-
-def test_ChandrasekharDynamicalFrictionForce_evaloutsideminrmaxr():
-    # Test that dynamical friction returns the expected force when evaluating
-    # outside of the [minr,maxr] range over which sigmar is interpolated:
-    # 0 at r < minr
-    # using sigmar(r) for r > maxr
-    from galpy.util import bovy_conversion
-    ro,vo= 8.,220.
-    # Parameters
-    GMs= 10.**9./bovy_conversion.mass_in_msol(vo,ro)
-    # Compute evolution with variable ln Lambda
-    sigmar= lambda r: 1./r
-    cdf= potential.ChandrasekharDynamicalFrictionForce(\
-        GMs=GMs,rhm=0.125,
-        dens=potential.MWPotential2014,sigmar=sigmar,
-        minr=0.5,maxr=2.)
-    # cdf 2 for checking r > maxr of cdf
-    cdf2= potential.ChandrasekharDynamicalFrictionForce(\
-        GMs=GMs,rhm=0.125,
-        dens=potential.MWPotential2014,sigmar=sigmar,
-        minr=0.5,maxr=4.)
-    v= [0.1,0.,0.]
-    # r < minr
-    assert numpy.fabs(cdf.Rforce(0.1,0.,v=v)) < 1e-16, 'potential.ChandrasekharDynamicalFrictionForce at r < minr not equal to zero'
-    assert numpy.fabs(cdf.zforce(0.1,0.,v=v)) < 1e-16, 'potential.ChandrasekharDynamicalFrictionForce at r < minr not equal to zero'
-    # r > maxr
-    assert numpy.fabs(cdf.Rforce(3.,0.,v=v)-cdf2.Rforce(3.,0.,v=v)) < 1e-10, 'potential.ChandrasekharDynamicalFrictionForce at r > maxr not as expected'
-    assert numpy.fabs(cdf.zforce(3.,0.,v=v)-cdf2.zforce(3.,0.,v=v)) < 1e-10, 'potential.ChandrasekharDynamicalFrictionForce at r > maxr not as expected'
-    return None
-
-def test_ChandrasekharDynamicalFrictionForce_pickling():
-    # Test that ChandrasekharDynamicalFrictionForce objects can/cannot be 
-    # pickled as expected
-    import pickle
-    from galpy.util import bovy_conversion
-    ro,vo= 8.,220.
-    # Parameters
-    GMs= 10.**9./bovy_conversion.mass_in_msol(vo,ro)
-    # sigmar internally computed, should be able to be pickled
-    # Compute evolution with variable ln Lambda
-    cdf= potential.ChandrasekharDynamicalFrictionForce(\
-        GMs=GMs,rhm=0.125,
-        dens=potential.MWPotential2014,
-        minr=0.5,maxr=2.)
-    pickled= pickle.dumps(cdf)
-    cdfu= pickle.loads(pickled)
-    # Test a few values
-    assert numpy.fabs(cdf.Rforce(1.,0.2,v=[1.,1.,0.])\
-                          -cdfu.Rforce(1.,0.2,v=[1.,1.,0.])) < 1e-10, 'Pickling of ChandrasekharDynamicalFrictionForce object does not work as expected'
-    assert numpy.fabs(cdf.zforce(2.,-0.2,v=[1.,1.,0.])\
-                          -cdfu.zforce(2.,-0.2,v=[1.,1.,0.])) < 1e-10, 'Pickling of ChandrasekharDynamicalFrictionForce object does not work as expected'
-    # Not providing dens = Logarithmic should also work
-    cdf= potential.ChandrasekharDynamicalFrictionForce(\
-        GMs=GMs,rhm=0.125,
-        minr=0.5,maxr=2.)
-    pickled= pickle.dumps(cdf)
-    cdfu= pickle.loads(pickled)
-    # Test a few values
-    assert numpy.fabs(cdf.Rforce(1.,0.2,v=[1.,1.,0.])\
-                          -cdfu.Rforce(1.,0.2,v=[1.,1.,0.])) < 1e-10, 'Pickling of ChandrasekharDynamicalFrictionForce object does not work as expected'
-    assert numpy.fabs(cdf.zforce(2.,-0.2,v=[1.,1.,0.])\
-                          -cdfu.zforce(2.,-0.2,v=[1.,1.,0.])) < 1e-10, 'Pickling of ChandrasekharDynamicalFrictionForce object does not work as expected'
-
-    # Providing sigmar as a lambda function gives AttributeError
-    sigmar= lambda r: 1./r
-    cdf= potential.ChandrasekharDynamicalFrictionForce(\
-        GMs=GMs,rhm=0.125,
-        dens=potential.MWPotential2014,sigmar=sigmar,
-        minr=0.5,maxr=2.)
-    if PY3:
-        with pytest.raises(AttributeError) as excinfo:
-            pickled= pickle.dumps(cdf)
-    else:
-        with pytest.raises(pickle.PicklingError) as excinfo:
-            pickled= pickle.dumps(cdf)
     return None
 
 def test_RingPotential_correctPotentialIntegral():
